basic sketch of an open source landing page
Even though we don't have a ton of open-source stuff, we do have some
stuff and we do use it, so seems logical to hype it on a page, if for no
other reason that candidates and other curios can see what's up.

This design just copies the team one and sucks, but maybe it'll get the
juices flowing.

* Would be nice to generate this somehow from our gemspecs or actual repos, but I don't know how to do that with Jekyll.
* Needs more imagery/color.  I really love Square's http://square.github.io/
* Do we show our forks of various things, too?